version 0.4.5-rc4
2009-02-26
	-文件已存在时自动重命令文件以防覆盖
	-加入音量控制功能
2009-02-21
	-实现面板上的来消息提示
2009-02-20
	-完成反安装程序
2009-02-19
	-更人性化的文件保存路径选择


git-svn-id: https://iptux.googlecode.com/svn/trunk@55 38dc79a6-912d-11dd-8943-ed8a20f67c6a

+++ src/my_entry.cpp
@@ -59,7 +59,7 @@ gboolean my_entry::QueryTooltip(GtkWidget * widget, gint x, gint y,
 void my_entry::InsertText(GtkEditable * editable, gchar * text, gint length,
 			  gint * position, gpointer data)
 {
-	ssize_t count;
+	gint count;
 
 	if (length == -1)
 		length = strlen(text);
